Moved SpawnInfo
