+++ core/src/com/week1/game/Model/SpawnInfo.java
@@ -1,4 +1,4 @@
-package com.week1.game.Renderer;
+package com.week1.game.Model;
 
 public class SpawnInfo {
